requirements.txt
cffi==1.14.3
pycparser==2.20
pycryptodome==3.9.8
six==1.15.0
